bring lemon to compile state

+++ src/xerror.h
@@ -17,6 +17,8 @@
 
 #pragma once
 
+#include <stddef.h>
+
 #include "../extern/cexception/CException.h"
 #include "lib/str.h"
 
@@ -79,7 +81,7 @@ const char *XerrorDescription(const xerror err);
 
 //display an error message to stderr in red font; does not log to xerror buffer
 //if line is > 0 then the message is prefixed with "line #:"
-void XerrorUser(const int line, const string msg, ...); 
+void XerrorUser(const size_t line, const string msg, ...); 
 
 //colours provided by @gon1332 at stackoverflow.com/questions/2616906/
 #ifdef COLOURS
